+++ reusable/__init__.py
@@ -6,13 +6,13 @@
 A python package with reusable functions and utility classes. 
 """
 __author__="Ohidur Rahman Bappy"
-__version__="0.0.8"
+__version__="0.0.9"
 __license__ = "MIT"
 __status__ = "Production"
 
 __all__=[
-    'AppConfig','Config','Namespace'
+    'AppConfig','Config','JsonCache'
 ]
 
-from .config import AppConfig,Config
-from .namespace import Namespace
+from .config import AppConfig,Config,JsonCache
+