+++ reusable/iterable_functions.py
@@ -34,3 +34,17 @@ def groupby_count(i, key=None, force_keys=None):
             counter[k] += 0
 
     return counter.items()
+
+
+def chunks(lst, n):
+    """Yield successive n-sized chunks from lst.
+    
+    Usage::
+
+        >>> from reusable.functions import chunks
+        >>> li=[1,3,5,6,8]
+        >>> for g in chunks(li,2):
+        ...     print(g)
+    """
+    for i in range(0, len(lst), n):
+        yield lst[i:i + n]